Fix releasechangelog.json path
Move generateMarkdown to core
